+++ Sources/BraintreeCore/BTAPIClient.m
@@ -272,7 +272,7 @@ NSString *const BTAPIClientErrorDomain = @"com.braintreepayments.BTAPIClientErro
 
     [self GET:@"v1/payment_methods"
              parameters:@{@"default_first": defaultFirstValue,
-                          @"session_id": self.metadata.sessionId}
+                          @"session_id": self.metadata.sessionID}
              completion:^(BTJSON * _Nullable body, __unused NSHTTPURLResponse * _Nullable response, NSError * _Nullable error) {
                  dispatch_async(dispatch_get_main_queue(), ^{
                      if (completion) {